refactor: Extract application exceptions into application/exceptions.py
Move ApplicationException, HandlerAlreadyRegistered, and UnsupportedHandler
to their own module. Keep error data classes in errors.py.
