+++ sharedkernel/application/__init__.py
@@ -1,9 +1,11 @@
-from . import commands, events, queries, services, validators
+from . import commands, errors, events, exceptions, queries, services, validators
 
 __all__ = [
     "commands",
+    "errors",
     "events",
+    "exceptions",
     "queries",
     "services",
-    "validators"
+    "validators",
 ]
